deleted vocab arg from decoder

+++ captioning_e3nn/decoder/decoder.py
@@ -346,7 +346,7 @@ class MyDecoderWithAttention(nn.Module):
         """Samples SMILES tockens for given  features (Greedy search).
         """
         h, c = self.init_hidden_state(features)
-
+        features = features.long()
         sampled_ids = []
         inputs = features.unsqueeze(1)
         for i in range(self.max_seg_length):
